// src/src/models/VendorProject/Vendor.tsx
export interface jobtypeinterface {
    gen_job_type: string;
    id: number;
    job_type_name: string;
}

export interface listvenderinterface{
    id: number 
    vendor_key: number 
    password: number 
    company_name: string
    add_datetime: number 
    email : string
    manager_name: string
    manager_role: number
    phone_number: number  
    affiliated: string 
    vendor_type: string
    location_detail: string
    note: string 
    vendor_level: number
    location_main_id : number
    vendor_project_id : number
    tambons_name_th : string
    amphures_name_th : string
    provinces_name_th : string
    zip_code : number
}